Conflicts:
	setup.py

+++ slapos/grid/SlapObject.py
@@ -43,6 +43,7 @@ from exception import BuildoutFailedError, WrongPermissionError, \
     PathDoesNotExistError
 from networkcache import download_network_cached, upload_network_cached
 import tarfile
+from watchdog import getWatchdogID
 
 REQUIRED_COMPUTER_PARTITION_PERMISSION = '0750'
 
@@ -55,7 +56,8 @@ class Software(object):
       shacache_key_file=None, shadir_cert_file=None, shadir_key_file=None,
       download_binary_cache_url=None, upload_binary_cache_url=None,
       download_binary_dir_url=None, upload_binary_dir_url=None,
-      binary_cache_url_blacklist = []):
+      download_from_binary_cache_url_blacklist = [],
+      upload_to_binary_cache_url_blacklist = []):
     """Initialisation of class parameters
     """
     self.url = url
@@ -78,7 +80,10 @@ class Software(object):
     self.upload_binary_cache_url = upload_binary_cache_url
     self.download_binary_dir_url = download_binary_dir_url
     self.upload_binary_dir_url = upload_binary_dir_url
-    self.binary_cache_url_blacklist = binary_cache_url_blacklist
+    self.download_from_binary_cache_url_blacklist = \
+        download_from_binary_cache_url_blacklist
+    self.upload_to_binary_cache_url_blacklist = \
+        upload_to_binary_cache_url_blacklist
 
   def install(self):
     """ Fetches binary cache if possible.
@@ -97,7 +102,7 @@ class Software(object):
             self.software_url_hash,
             tarpath, self.logger,
             self.signature_certificate_list,
-            self.binary_cache_url_blacklist):
+            self.download_from_binary_cache_url_blacklist):
       tar = tarfile.open(tarpath)
       try:
         self.logger.info("Extracting archive of cached software release...")
@@ -106,9 +111,18 @@ class Software(object):
         tar.close()
     else:
       self._install_from_buildout()
+
+      # Upload to binary cache if possible
+      blacklisted = False
+      for url in self.upload_to_binary_cache_url_blacklist:
+        if self.url.startswith(url):
+          blacklisted = True
+          self.logger.debug("Can't download from binary cache: "
+              "Software Release URL is blacklisted.")
       if (self.software_root and self.url and self.software_url_hash \
                              and self.upload_binary_cache_url \
-                             and self.upload_binary_dir_url):
+                             and self.upload_binary_dir_url \
+                             and not blacklisted):
         self.logger.info("Creating archive of software release...")
         tar = tarfile.open(tarpath, "w:gz")
         try:
@@ -224,6 +238,7 @@ class Partition(object):
     self.software_path = software_path
     self.instance_path = instance_path
     self.run_path = os.path.join(self.instance_path, 'etc', 'run')
+    self.service_path = os.path.join(self.instance_path, 'etc', 'service')
     self.supervisord_partition_configuration_path = \
         supervisord_partition_configuration_path
     self.supervisord_socket = supervisord_socket
@@ -263,6 +278,26 @@ class Partition(object):
     gid = stat_info.st_gid
     return (uid, gid)
 
+  def addServiceToGroup(self, partition_id,
+                        runner_list, path, extension = ''):
+    uid, gid = self.getUserGroupId()
+    program_partition_template = pkg_resources.resource_stream(__name__,
+            'templates/program_partition_supervisord.conf.in').read()
+    for runner in runner_list:
+      self.partition_supervisor_configuration += '\n' + \
+          program_partition_template % dict(
+        program_id='_'.join([partition_id, runner]),
+        program_directory=self.instance_path,
+        program_command=os.path.join(path, runner),
+        program_name=runner+extension,
+        instance_path=self.instance_path,
+        user_id=uid,
+        group_id=gid,
+        # As supervisord has no environment to inherit setup minimalistic one
+        HOME=pwd.getpwuid(uid).pw_dir,
+        USER=pwd.getpwuid(uid).pw_name,
+      )
+
   def install(self):
     """ Creates configuration file from template in software_path, then
     installs the software partition with the help of buildout
@@ -285,7 +320,10 @@ class Partition(object):
     os.environ = utils.getCleanEnvironment(pwd.getpwuid(
       instance_stat_info.st_uid).pw_dir)
     # Generates buildout part from template
-    template_location = os.path.join(self.software_path, 'template.cfg')
+    template_location = os.path.join(self.software_path, 'instance.cfg')
+    # Backward compatibility: "instance.cfg" file was named "template.cfg".
+    if not os.path.exists(template_location):
+      template_location = os.path.join(self.software_path, 'template.cfg')
     config_location = os.path.join(self.instance_path, 'buildout.cfg')
     self.logger.debug("Copying %r to %r" % (template_location, config_location))
     try:
@@ -368,42 +406,35 @@ class Partition(object):
     self.logger.info("Generating supervisord config file from template...")
     # check if CP/etc/run exists and it is a directory
     # iterate over each file in CP/etc/run
+    # iterate over each file in CP/etc/service adding WatchdogID to their name
     # if at least one is not 0750 raise -- partition has something funny
     runner_list = []
+    service_list = []
     if os.path.exists(self.run_path):
       if os.path.isdir(self.run_path):
         runner_list = os.listdir(self.run_path)
-    if len(runner_list) == 0:
-      self.logger.warning('No runners found for partition %r' %
+    if os.path.exists(self.service_path):
+      if os.path.isdir(self.service_path):
+        service_list = os.listdir(self.service_path)
+    if len(runner_list) == 0 and len(service_list) == 0:
+      self.logger.warning('No runners nor services found for partition %r' %
           self.partition_id)
       if os.path.exists(self.supervisord_partition_configuration_path):
         os.unlink(self.supervisord_partition_configuration_path)
     else:
       partition_id = self.computer_partition.getId()
-      program_partition_template = pkg_resources.resource_stream(__name__,
-          'templates/program_partition_supervisord.conf.in').read()
       group_partition_template = pkg_resources.resource_stream(__name__,
           'templates/group_partition_supervisord.conf.in').read()
-      partition_supervisor_configuration = group_partition_template % dict(
+      self.partition_supervisor_configuration = group_partition_template % dict(
           instance_id=partition_id,
           program_list=','.join(['_'.join([partition_id, runner])
-            for runner in runner_list]))
-      for runner in runner_list:
-        partition_supervisor_configuration += '\n' + \
-            program_partition_template % dict(
-          program_id='_'.join([partition_id, runner]),
-          program_directory=self.instance_path,
-          program_command=os.path.join(self.run_path, runner),
-          program_name=runner,
-          instance_path=self.instance_path,
-          user_id=uid,
-          group_id=gid,
-          # As supervisord has no environment to inherit setup minimalistic one
-          HOME=pwd.getpwuid(uid).pw_dir,
-          USER=pwd.getpwuid(uid).pw_name,
-        )
+            for runner in runner_list+service_list]))
+      # Same method to add to service and run 
+      self.addServiceToGroup(partition_id, runner_list,self.run_path)
+      self.addServiceToGroup(partition_id, service_list,self.service_path,
+                             extension=getWatchdogID())
       utils.updateFile(self.supervisord_partition_configuration_path,
-          partition_supervisor_configuration)
+          self.partition_supervisor_configuration)
     self.updateSupervisor()
 
   def start(self):
